# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Fixture {
  id: number;
  homeTeam: string;
  awayTeam: string;
  league: string;
  country: string;
  kickoff: string;
  status: string;
}

export interface TodayFixturesResponse {
  fixtures: Fixture[];
  date: string;
  fetchedAt: string;
  nextRefreshAt: string;
  stale: boolean;
  /** @nullable */
  warning: string | null;
}

export type LiveFixture = Fixture & {
  minute: number;
  homeScore: number;
  awayScore: number;
};

export interface LiveFixturesResponse {
  fixtures: LiveFixture[];
  fetchedAt: string;
  nextRefreshAt: string;
  stale: boolean;
  /** @nullable */
  warning: string | null;
}

export interface FixturesUnavailable {
  error: string;
  warning: string;
}

export interface VapidPublicKeyResponse {
  publicKey: string;
}

export interface PushSubscriptionKeys {
  p256dh: string;
  auth: string;
}

export interface PushSubscription {
  endpoint: string;
  /** @nullable */
  expirationTime?: number | null;
  keys: PushSubscriptionKeys;
}

export interface FavoriteIdsRequest {
  fixtureIds: number[];
}

export interface SuccessResponse {
  ok: boolean;
}