# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: fixtures
    description: Real football fixtures with simulated analysis data
  - name: push
    description: Web Push subscriptions and favorite fixture monitoring
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /fixtures/today:
    get:
      operationId: getTodayFixtures
      tags: [fixtures]
      summary: Get today's football fixtures
      description: Returns today's fixtures from API-Football. The server refreshes the upstream data at most once every 2 hours and returns the last valid response when an upstream refresh fails.
      responses:
        "200":
          description: Today's fixtures and cache status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodayFixturesResponse"
        "503":
          description: No real fixture data is available yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FixturesUnavailable"
  /fixtures/live:
    get:
      operationId: getLiveFixtures
      tags: [fixtures]
      summary: Get genuinely live football fixtures
      description: Returns only fixtures currently in progress from API-Football. The upstream list is refreshed at most once every 20 minutes.
      responses:
        "200":
          description: Current live fixtures and cache status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LiveFixturesResponse"
        "503":
          description: No valid live-fixture response is available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FixturesUnavailable"
  /vapid-public-key:
    get:
      operationId: getVapidPublicKey
      tags: [push]
      summary: Get the public VAPID key
      responses:
        "200":
          description: Public key used by PushManager
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VapidPublicKeyResponse"
  /subscribe:
    post:
      operationId: subscribePush
      tags: [push]
      summary: Store a browser push subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscription"
      responses:
        "200":
          description: Subscription stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /favorites:
    post:
      operationId: updateFavoriteIds
      tags: [push]
      summary: Store fixture IDs that should be monitored
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteIdsRequest"
      responses:
        "200":
          description: Favorite IDs stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Fixture:
      type: object
      properties:
        id:
          type: integer
        homeTeam:
          type: string
        awayTeam:
          type: string
        league:
          type: string
        country:
          type: string
        kickoff:
          type: string
          format: date-time
        status:
          type: string
      required:
        - id
        - homeTeam
        - awayTeam
        - league
        - country
        - kickoff
        - status
    TodayFixturesResponse:
      type: object
      properties:
        fixtures:
          type: array
          items:
            $ref: "#/components/schemas/Fixture"
        date:
          type: string
          format: date
        fetchedAt:
          type: string
          format: date-time
        nextRefreshAt:
          type: string
          format: date-time
        stale:
          type: boolean
        warning:
          type: ["string", "null"]
      required:
        - fixtures
        - date
        - fetchedAt
        - nextRefreshAt
        - stale
        - warning
    LiveFixture:
      allOf:
        - $ref: "#/components/schemas/Fixture"
        - type: object
          properties:
            minute:
              type: integer
            homeScore:
              type: integer
            awayScore:
              type: integer
          required:
            - minute
            - homeScore
            - awayScore
    LiveFixturesResponse:
      type: object
      properties:
        fixtures:
          type: array
          items:
            $ref: "#/components/schemas/LiveFixture"
        fetchedAt:
          type: string
          format: date-time
        nextRefreshAt:
          type: string
          format: date-time
        stale:
          type: boolean
        warning:
          type: ["string", "null"]
      required:
        - fixtures
        - fetchedAt
        - nextRefreshAt
        - stale
        - warning
    FixturesUnavailable:
      type: object
      properties:
        error:
          type: string
        warning:
          type: string
      required:
        - error
        - warning
    VapidPublicKeyResponse:
      type: object
      properties:
        publicKey:
          type: string
      required:
        - publicKey
    PushSubscriptionKeys:
      type: object
      properties:
        p256dh:
          type: string
        auth:
          type: string
      required:
        - p256dh
        - auth
    PushSubscription:
      type: object
      properties:
        endpoint:
          type: string
          format: uri
        expirationTime:
          type: ["number", "null"]
        keys:
          $ref: "#/components/schemas/PushSubscriptionKeys"
      required:
        - endpoint
        - keys
    FavoriteIdsRequest:
      type: object
      properties:
        fixtureIds:
          type: array
          items:
            type: integer
      required:
        - fixtureIds
    SuccessResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok