# lib/api-zod/src/generated/types/pushSubscription.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PushSubscriptionKeys } from './pushSubscriptionKeys';

export interface PushSubscription {
  endpoint: string;
  /** @nullable */
  expirationTime?: number | null;
  keys: PushSubscriptionKeys;
}